CORE-6341 Move NotificationMessage model
