+++ ui/de-lib/src/main/java/org/iplantc/de/desktop/client/views/windows/NotificationWindow.java
@@ -36,15 +36,6 @@ import java.util.List;
  */
 public class NotificationWindow extends IplantWindowBase {
 
-    private class NotificationKeyProvider implements ModelKeyProvider<NotificationMessage> {
-
-        @Override
-        public String getKey(NotificationMessage item) {
-            return item.getId();
-        }
-
-    }
-
     private static CheckBoxSelectionModel<NotificationMessage> checkBoxModel;
     private final IplantDisplayStrings displayStrings;
     private NotificationView.Presenter presenter;
